[CLN] Dispatch (some) Frame ops to Series, avoiding _data.eval (#22019)
* avoid casting to object dtype in mixed-type frames

* Dispatch to Series ops in _combine_match_columns

* comment

* docstring

* flake8 fixup

* dont bother with try_cast_result

* revert non-central change

* simplify

* revert try_cast_results

* revert non-central changes

* Fixup typo syntaxerror

* simplify assertion

* use dispatch_to_series in combine_match_columns

* Pass unwrapped op where appropriate

* catch correct error

* whatsnew note

* comment

* whatsnew section

* remove unnecessary tester

* doc fixup

+++ pandas/tests/frame/test_axis_select_reindex.py
@@ -721,7 +721,7 @@ class TestDataFrameSelectReindex(TestData):
 
         result = df1 - df1.mean()
         expected = df2 - df2.mean()
-        assert_frame_equal(result, expected)
+        assert_frame_equal(result.astype('f8'), expected)
 
     def test_align_multiindex(self):
         # GH 10665
